+++ app/views/mappings/automap_popup.js
@@ -1,0 +1,2 @@
+$("#mappings-popup-container").html("<%= escape_javascript(render("automap_popup")) %>");
+showContourModal();